#14: Remove inner joins, create scan controller, rename failure to error
