+++ App/app/src/main/res/values/strings.xml
@@ -16,13 +16,13 @@
     <string name="menu_sign_out">Sign Out</string>
     <string name="drawer_open" translatable="false">Drawer opened</string>
     <string name="drawer_close" translatable="false">Drawer closed</string>
-    <string name="scanner_permissions_failure">You need to grant camera and storage permissions</string>
-    <string name="scanner_initialization_failure">Scanner initialization failed</string>
+    <string name="scanner_permissions_error">You need to grant camera and storage permissions</string>
+    <string name="scanner_initialization_error">Scanner initialization failed</string>
     <string name="scan">Scan</string>
-    <string name="scanner_general_failure">Something bad happened to scanner</string>
-    <string name="scanner_success">Scanner success</string>
+    <string name="scanner_general_error">Something bad happened to scanner</string>
+    <string name="scanner_success">You have added a new scan</string>
     <string name="server_connection_error">Server connection error</string>
-    <string name="scan_upload_failure">Something went wrong with uploading the scan</string>
+    <string name="scan_upload_error">Something went wrong with uploading the scan</string>
     <string name="save">save</string>
     <string name="cancel">cancel</string>
     <string name="try_again">Something went wrong. Please try again</string>
